Added points and lines to O.D. 89
Added new type of representation of photos to Oude Delft 89 in the form of point and line geojsons. The geojson consists of a point and line group, with information about the photograph being added to the properties of the point. These new geojsons have been added to the 11th and 12th slides of the building.